Table: Delete all client references to groupToNestedTableV2 flag, functionally GAing the transformation
