+++ public/app/features/transformers/docs/content.ts
@@ -711,8 +711,6 @@ Use this transformation to construct a matrix by specifying fields from your que
 
   #### Display options
 
-> **Note:** Display options are in public preview. To try out the new editor for this transformation, enable the \`groupToNestedTableV2\` feature toggle. To try out nested field overrides, enable \`nestedFramesFieldOverrides\`.
-
   | Option                            | Description                                                                                                                |
   | --------------------------------- | -------------------------------------------------------------------------------------------------------------------------- |
   | Show field names in nested tables | Show or hide the column headers inside each nested table. On by default                                                   |
